Add user id in Todos and validation on web services

+++ src/main/java/com/ahmad/springboot/todos/TodoService.java
@@ -4,6 +4,7 @@ import com.ahmad.springboot.error.ConflictException;
 import com.ahmad.springboot.error.NotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.NotAcceptableStatusException;
 
 import java.util.List;
 import java.util.NoSuchElementException;
@@ -14,13 +15,12 @@ public class TodoService {
     @Autowired
     private TodoRepository todoRepository;
 
-    /**
-     * Get all todos
-     *
-     * @return List<Todo>
-     */
     public List<Todo> findAll() {
         return todoRepository.findAll();
+    }
+
+    public List<Todo> findByUser(String id) {
+        return todoRepository.findByUserId(id);
     }
 
     public Todo getById(String id) {
@@ -31,13 +31,25 @@ public class TodoService {
         }
     }
 
+    public Todo getByUserIdAndId(String userId, String id) {
+
+        if (!todoRepository.existsByUserIdAndId(userId, id))
+            throw new NotFoundException(String.format("No Record with the id [%s] was found in the database", id));
+
+        return todoRepository.findByIdAndUserId(id, userId);
+    }
+
     public Todo save(Todo todo) {
         if (todoRepository.findByTitle(todo.getTitle()) != null)
             throw new ConflictException("Another record with the same title exists");
+
         return todoRepository.insert(todo);
     }
 
-    public void delete(String id) {
+    public void delete(String userId, String id) {
+        if (!todoRepository.existsByUserIdAndId(userId, id))
+            throw new NotAcceptableStatusException("Id is not correct");
+
         todoRepository.deleteById(id);
     }
 
@@ -45,8 +57,13 @@ public class TodoService {
         if (!todoRepository.existsById(todo.getId()))
             throw new NotFoundException(String.format("No Record with the id [%s] was found in the database", todo.getId()));
 
+        if (!todoRepository.existsByUserIdAndId(todo.getUserId(), todo.getId()))
+            throw new NotAcceptableStatusException("User id is not correct");
+
         if (todoRepository.findByTitleAndIdNot(todo.getTitle(), todo.getId()) != null)
             throw new ConflictException("Another record with the same title exists");
+
         return todoRepository.save(todo);
     }
+
 }
